refactor: split reward and referral program actions into per-endpoint files
Both action files had grown into grab-bags. Follow the existing
createApiReward/ convention: one file per endpoint plus a shared utils
file and a barrel.

Also drops the duplicated reward loader in requireLinkableReward, which
now delegates to requireApiReward, and replaces the nested ternaries in
mergeProgramUpdate with a named patchField helper.
