+++ server/src/internal/rewards/actions/rewardCrud/index.ts
@@ -1,0 +1,4 @@
+export { requireApiReward, toApiRewardResponse } from "./apiRewardUtils.js";
+export { deleteApiReward } from "./deleteApiReward.js";
+export { getApiRewardById } from "./getApiRewardById.js";
+export { updateApiReward } from "./updateApiReward.js";
